server/requirements.txt
Flask==3.1.3
Flask-CORS==6.0.2
stripe==14.3.0
python-dotenv==1.2.1
gunicorn==21.2.0
Werkzeug==3.1.6
